requirements.txt
numpy
pandas
scikit-learn
xgboost
praat-parselmouth
joblib
streamlit
streamlit-mic-recorder